requirements.txt
absl-py==2.1.0
blessed==1.20.0
bpython==0.24
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
curtsies==0.4.2
cwcwidth==0.1.9
exceptiongroup==1.2.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud==0.34.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-kms==2.21.3
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.18.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pytest==8.1.1
python-gnupg==0.5.2
pyxdg==0.28
requests==2.31.0
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==2.2.1
wcwidth==0.2.13
